fix: debuging the export option

+++ src/FileUtils.java
@@ -1,19 +1,39 @@
 import java.io.*;
+import java.net.URISyntaxException;
 
 public class FileUtils {
     public boolean exportNote(Note note) {
-        File exportedNote = new File("/export/" + note.getTitle() + ".txt");
-
+        File exportedNote;
+        new File("export").mkdir();
         try {
-            exportedNote.delete();
-            exportedNote.createNewFile();
-        } catch (IOException e) {
+            String jarPath = getClass().getProtectionDomain().getCodeSource()
+                    .getLocation().toURI().getPath();
+
+            String jarName = jarPath.substring(jarPath.lastIndexOf("/") + 1);
+
+            exportedNote = new File(jarPath.replace(jarName,"") + "export/" + note.getTitle() +
+                    ".txt");
+
+        } catch (URISyntaxException e) {
+            e.printStackTrace();
             return false;
         }
 
         try {
-            DataOutputStream out = new DataOutputStream(new FileOutputStream(exportedNote));
-            out.writeUTF(note.getDisplayFormat());
+            if (exportedNote.exists()) {
+                exportedNote.delete();
+            }
+            exportedNote.createNewFile();
+        } catch (IOException e) {
+            e.printStackTrace();
+            return false;
+        }
+
+        try {
+            FileWriter out = new FileWriter(exportedNote);
+            out.write(note.getDisplayFormat());
+            out.flush();
+            out.close();
             return true;
         } catch (IOException fileNotFoundException) {
             fileNotFoundException.printStackTrace();
